requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pika
pydantic
python-multipart
python-dotenv